[UPD] tests

+++ src/MusicBox.sol
@@ -11,21 +11,20 @@ import {IMusicBox} from "src/interfaces/MusicBox/IMusicBox.sol";
 contract MusicBox is Base721, IMusicBox {
     uint256 public constant MAX_SUPPLY = 3333;
 
-    mapping(uint tokenId => MusicBoxLevel) public tokenLevel;
+    mapping(uint256 tokenId => MusicBoxLevel) public tokenLevel;
 
     constructor(string memory _name, string memory _symbol, string memory _contractURI, string memory _baseURI)
         Base721(_name, _symbol, _contractURI, _baseURI)
     {}
 
-    function mintFromSantuary(address _to, MusicBoxLevel musicBoxLevel, uint _amount) external onlyOwner {
+    function mintFromSantuary(address _to, MusicBoxLevel musicBoxLevel, uint256 _amount) external onlyOwner {
         unchecked {
             for (uint256 i = 0; i < _amount; i++) {
-                uint newId = _getTokenIdAndIncrement();
+                uint256 newId = _getTokenIdAndIncrement();
                 userMinted[_to] += 1;
                 tokenLevel[newId] = musicBoxLevel;
                 _safeMint(_to, newId);
             }
         }
     }
-
 }
